refactor: Remove the extra serac from the test files and names
