+++ src/serac/infrastructure/tests/CMakeLists.txt
@@ -8,9 +8,9 @@ set(test_dependencies gtest serac_physics serac_state)
 blt_list_append( TO test_dependencies ELEMENTS caliper adiak IF ${SERAC_ENABLE_PROFILING} )
 
 set(infrastructure_tests
-    serac_error_handling.cpp
-    serac_input.cpp
-    serac_profiling.cpp)
+    error_handling.cpp
+    input.cpp
+    profiling.cpp)
 
 serac_add_tests( SOURCES ${infrastructure_tests}
                  DEPENDS_ON ${test_dependencies})
