+++ src/serac/mesh/tests/CMakeLists.txt
@@ -8,7 +8,7 @@ set(test_dependencies gtest serac_numerics serac_mesh serac_boundary_conditions 
 blt_list_append( TO test_dependencies ELEMENTS caliper adiak IF ${SERAC_ENABLE_PROFILING} )
 
 set(mesh_parallel_tests
-    serac_mesh_test.cpp
+    mesh_test.cpp
     )
 
 serac_add_tests( SOURCES ${mesh_parallel_tests}
