+++ src/serac/physics/benchmarks/CMakeLists.txt
@@ -8,11 +8,11 @@ set(benchmark_dependencies serac_physics)
 blt_list_append(TO benchmark_dependencies ELEMENTS caliper adiak IF ${SERAC_ENABLE_PROFILING})
 blt_list_append(TO benchmark_dependencies ELEMENTS mpi IF ${ENABLE_MPI})
 
-blt_add_executable(NAME benchmark_serac_thermal_functional
-                   SOURCES benchmark_serac_thermal_functional.cpp
+blt_add_executable(NAME benchmark_thermal
+                   SOURCES benchmark_thermal.cpp
                    DEPENDS_ON ${benchmark_dependencies}
                    OUTPUT_DIR ${PROJECT_BINARY_DIR}/benchmarks
                    FOLDER serac/benchmarks)
 
-target_include_directories(benchmark_serac_thermal_functional
+target_include_directories(benchmark_thermal
                            SYSTEM PRIVATE ${adiak_INCLUDE_DIRS})
