+++ src/serac/physics/boundary_conditions/tests/CMakeLists.txt
@@ -8,7 +8,7 @@ set(test_dependencies serac_physics test_utils)
 blt_list_append( TO test_dependencies ELEMENTS caliper adiak IF ${SERAC_ENABLE_PROFILING} )
 
 set(boundary_cond_tests
-    serac_boundary_cond.cpp)
+    boundary_cond.cpp)
 
 serac_add_tests( SOURCES ${boundary_cond_tests}
                  DEPENDS_ON ${test_dependencies}
