+++ src/serac/physics/tests/solid_legacy_adjoint.cpp
