+++ src/tests/cuda_smoketest.cpp
