+++ src/tests/cuda_smoketest_kernel.cpp
